Move files under snapshots
